Fix remaining shadowing, duplicate @Override, type conversion, and access errors
- Fully qualify org.cameracontroller.PngjBadSignature.values in 3 Verify*Activity files
- Remove duplicate PngjBadCrcException import in TouchLessIDSDK$values.java
- Stub zzic methods in defpackage/a.java (package-private GMS class)
- Remove 46 duplicate @Override annotations across 22 response files
- Fix type conversions in arid/Action.java (String→byte[], int→byte, remove final)
- Fix lossy byte cast in arid/a/valueOf.java

+++ app/src/main/java/arid/a/valueOf.java
@@ -158,7 +158,7 @@ public abstract class valueOf extends OutputStream implements AutoCloseable {
         int i5 = 3 - (b2 * 4);
         byte[] bArr2 = new byte[i4];
         if (bArr == null) {
-            byte b3 = r9;
+            byte b3 = (byte) r9;
             i3 = 0;
             int i6 = i5;
             int i7 = (i5 + (-b3)) - 52;
